GatorGabbeler/server/requirements.txt
# requirements in order to run the GatorGabbeler server
fastapi
uvicorn[standard]
python-dotenv
httpx

# RAG dependencies - use with Python 3.11 or 3.12 for pre-built wheels DO NOT USE IF PY 3.14 TOO MUCH TIME 
langchain
langchain-openai
langchain-community
pypdf
chromadb
tiktoken